pyproject.toml
[project]
name = "mstodo-openclaw-skill"
version = "2.0.0"
description = "Microsoft To Do — OpenClaw Skill (CLI-based)"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "aiohttp==3.9.1",
    "pydantic==2.5.2",
    "pytz==2023.3",
    "httpx==0.27.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
mqtt = [
    "paho-mqtt>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "todo", "utils"]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
